backend/requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# CORS
django-cors-headers==4.3.1

# Image handling
Pillow==10.2.0

# Phone number validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# API documentation
drf-spectacular==0.27.0

# Celery for async tasks
celery==5.3.6
redis==5.0.1

# Security
django-ratelimit==4.1.0

# Utilities
python-decouple==3.8
python-dateutil==2.8.2
